Fix for JCS-106: Bad comparison of int with 0x7fffffffffffffffL
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/jcs/trunk@1517262 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/jcs/auxiliary/remote/server/RemoteCacheServer.java
@@ -1603,7 +1603,7 @@ public class RemoteCacheServer<K extends Serializable, V extends Serializable>
     private long nextListenerId()
     {
         long id = 0;
-        if ( listenerId[0] == Long.MAX_VALUE )
+        if ( listenerId[0] == Integer.MAX_VALUE )
         {
             synchronized ( listenerId )
             {
